requirements.txt
Django==2.2.7
django-rest-framework==0.1.0
djangorestframework==3.10.3
gunicorn==20.0.0
pytz==2019.3
sqlparse==0.3.0
whitenoise==4.1.4
